Improve board colors and make cells a 2D grid
