+++ src/__init__.py
@@ -1,0 +1,3 @@
+from .cell import *
+from .game import *
+from .display import *
